docs: fix mkdocs building
