+++ doc/mkdocs/requirements.txt
@@ -1,16 +1,14 @@
-mkdocs-material
-mkdocs-jupyter
-ipython_genutils
-python-markdown-math
-mkdocs-static-i18n
-mkdocstrings
-mkdocstrings-python-legacy
-pytkdocs[numpy-style]
-mkdocs-simple-hooks
-mkdocs-awesome-pages-plugin
-mkdocs-git-revision-date-plugin
-GitPython
-
+GitPython==3.1.27
+mkdocs-material==8.4.2
+mkdocs-jupyter==0.21.0
+ipython_genutils==0.2.0
+python-markdown-math==0.8
+mkdocs-static-i18n==0.46
+mkdocstrings[python]==0.19.0
+mkdocs-simple-hooks==0.1.5
+mkdocs-awesome-pages-plugin==2.8.0
+mkdocs-git-revision-date-plugin==0.3.2
+mkdocs-gen-files==0.4.0
 
 # for FATE api generate
 peewee
